+++ test/utils/countDeprecatedTimezones.js
@@ -1,0 +1,5 @@
+import { getDeprecatedTimezones } from './getDeprecatedTimezones';
+
+export const countDeprecatedTimezones = (timezones) => {
+  return Object.keys(getDeprecatedTimezones(timezones)).length;
+}