+++ test/utils/index.js
@@ -1,0 +1,2 @@
+export * from './countDeprecatedTimezones';
+export * from './getDeprecatedTimezones';