+++ types/index.d.ts
@@ -17,24 +17,29 @@ export interface Timezone {
   dstOffset: number;
   dstOffsetStr: string;
   aliasOf: string | null;
+  deprecated: boolean | undefined;
 }
 
-export function getCountry(id: CountryCode): Country;
-export function getCountry(id: string): Country | null;
+export interface Options {
+  deprecated: boolean;
+}
+
+export function getCountry(id: CountryCode, options?: Options): Country;
+export function getCountry(id: string, options?: Options): Country | null;
 
 export function getTimezone(name: TimezoneName): Timezone;
 export function getTimezone(name: string): Timezone | null;
 
-export function getAllCountries(): Record<CountryCode, Country>;
+export function getAllCountries(options?: Options): Record<CountryCode, Country>;
 
-export function getAllTimezones(): Record<TimezoneName, Timezone>;
+export function getAllTimezones(options?: Options): Record<TimezoneName, Timezone>;
 
-export function getTimezonesForCountry(id: CountryCode): Timezone[];
-export function getTimezonesForCountry(id: string): Timezone[] | null;
+export function getTimezonesForCountry(id: CountryCode, options?: Options): Timezone[];
+export function getTimezonesForCountry(id: string, options?: Options): Timezone[] | null;
 
-export function getCountriesForTimezone(name: TimezoneName): Country[];
-export function getCountriesForTimezone(name: string): Country[];
-export function getCountriesForTimezone(name: string): Country[] | null;
+export function getCountriesForTimezone(name: TimezoneName, options?: Options): Country[];
+export function getCountriesForTimezone(name: string, options?: Options): Country[];
+export function getCountriesForTimezone(name: string, options?: Options): Country[] | null;
 
-export function getCountryForTimezone(name: TimezoneName): Country;
-export function getCountryForTimezone(name: string): Country | null;
+export function getCountryForTimezone(name: TimezoneName, options?: Options): Country;
+export function getCountryForTimezone(name: string, options?: Options): Country | null;